refactor indexing service and remove semantics descriptions
